Added support for deterministic language detection and language prioritization.
Added support for deterministic language detection, e.g. that the language same text can be detected with the same probability. Previously this wasn’t possible since the static configuration state of the DetectorFactory always interfered with the construction of the Detector instance, effectively overriding the configured state.

We also support to restrict the language detection to a given set of languages with weights via the new languagePriorities property.

We now construct the Detector instance ourselves which can be configured accordingly. We also allow a Detector instance to be reused by manually resetting the Detector state via the reset(Detector) method.