XD-3056 - Add video capturing source module.
This module supports capturing frames from video, which can originate from either a camera or file. A message with the JPEG-encoded image data as payload is emitted into the stream For each grabbed frame of the input video. The main use case is to allow to ingest and decode video data into a sequence of images, amenable to later processing, i.e. object detection with computer-vision algorithms.

Original pull request: #14.